\name{border.functions}
\alias{border.functions}
\docType{data}
\title{
List Of Border Functions
}
\description{
A list of functions which automatically allocate borders to nodes representing HPO terms in plots generated by \code{\link{hpo.graph}}.
}
\format{
  The format is:
List of 2
 $ get.grey.borders:function (hpo.terms, plotting.context)  
 $ get.no.borders  :function (hpo.terms, plotting.context)  
}
\keyword{datasets}
\keyword{border}
