% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{calibrate.sizes}
\alias{calibrate.sizes}
\title{Function to scale sizes of terms between two given limits}
\usage{
calibrate.sizes(x, high, low)
}
\arguments{
\item{x}{Numeric vector of term relative sizes named by term}

\item{high}{Numeric vector of largest size}

\item{low}{Numeric vector of smallest size}
}
\value{
Numeric vector
}
\description{
Function to scale sizes of terms between two given limits
}
\examples{
calibrate.sizes(c("HP:0000001"=10, "HP:0000006"=5), high=3, low=1)
}

