\name{colouring.functions}
\alias{colouring.functions}
\docType{data}
\title{
List Of Colouring Functions
}
\description{
A list of functions which automatically generate colours for nodes representing HPO terms in plots generated by \code{\link{hpo.graph}}.
}
\format{
  The format is:
List of 4
 $ get.patient.based.colours    :function (hpo.terms, plotting.context)  
 $ get.information.based.colours:function (hpo.terms, plotting.context, colourPalette = colorRampPalette(c("Yellow", 
    "Green", "#0099FF"))(10))  
 $ get.frequency.based.colours  :function (hpo.terms, plotting.context, colour.func = NULL)  
 $ get.white.nodes              :function (hpo.terms, plotting.context)  
}
\keyword{datasets}
\keyword{colour}
