% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{exclude.branch}
\alias{exclude.branch}
\title{Exclude terms descending from particular term from a character vector of terms}
\usage{
exclude.branch(hpo.terms, branch.root, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{branch.root}{HPO term whose descendants should be excluded}

\item{terms}{Character vector of HPO terms}
}
\value{
Character vector of terms
}
\description{
Exclude terms descending from particular term from a character vector of terms
}

