% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.case.based.colours}
\alias{get.case.based.colours}
\title{Function to set colours of HPO nodes in plot to distinguish terms belonging to different sets of phenotypes}
\usage{
get.case.based.colours(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector of colours, named by term
}
\description{
Function to set colours of HPO nodes in plot to distinguish terms belonging to different sets of phenotypes
}

