% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.case.based.labels}
\alias{get.case.based.labels}
\title{Function to label HPO nodes in plot to indicate to which phenotypes each of the terms belong}
\usage{
get.case.based.labels(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector of colours, named by term
}
\description{
Function to label HPO nodes in plot to indicate to which phenotypes each of the terms belong
}

