% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get.case.term.matrix}
\alias{get.case.term.matrix}
\title{Get a matrix with columns of hpo terms and rows of patients,}
\usage{
get.case.term.matrix(hpo.phenotypes, columns = NULL)
}
\arguments{
\item{hpo.phenotypes}{List of character vectors of HPO terms. Result includes only terms which are explicitly present in the list items, so if you wish the result to include even terms which are implicitly present, lapply \code{\link{get.ancestors}} to the argument before passing it to this function}

\item{columns}{Force result to have these exact columns, entering F for terms which aren't present}
}
\value{
Logical matrix - entry for a patient/hpo term = T if the patient has the term and F otherwise.
}
\description{
Get a matrix with columns of hpo terms and rows of patients,
}
\examples{
get.case.term.matrix(list(Patient1=c("HP:0001873")))
}

