% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.frequency.based.colours}
\alias{get.frequency.based.colours}
\title{Function to colour HPO nodes in plot with colours based on frequency with which terms appear in phenotypes}
\usage{
get.frequency.based.colours(hpo.terms, terms, plotting.context,
  colour.func = NULL)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}

\item{colour.func}{Function capable of returning a set of colours, given the number of colours it needs to return}
}
\value{
Character vector of colours, named by term
}
\description{
Function to colour HPO nodes in plot with colours based on frequency with which terms appear in phenotypes
}

