% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.informative.node.labels}
\alias{get.informative.node.labels}
\title{Function to label HPO nodes in plot with node description and information content}
\usage{
get.informative.node.labels(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector of labels, named by term
}
\description{
Function to label HPO nodes in plot with node description and information content
}

