% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get.mpo.to.hpo}
\alias{get.mpo.to.hpo}
\title{Get MPO to HPO R-Object}
\usage{
get.mpo.to.hpo(hpo.terms, cross.species.file)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{cross.species.file}{cross species obo file, downloadable from http://compbio.charite.de/hudson/? website}
}
\value{
List of HPO terms per MPO term
}
\description{
Get MPO to HPO R-Object
}

