% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.no.borders}
\alias{get.no.borders}
\title{Function to border HPO nodes in plot with transparent borders}
\usage{
get.no.borders(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector of sizes, named by term
}
\description{
Function to border HPO nodes in plot with transparent borders
}

