% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.shortened.names}
\alias{get.shortened.names}
\title{Get human readable, shortened (where possible) HPO term names}
\usage{
get.shortened.names(hpo.terms, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}
}
\value{
Character vector
}
\description{
Get human readable, shortened (where possible) HPO term names
}
\examples{
data(hpo.terms)
get.shortened.names(hpo.terms, c("HP:0001873", "HP:0011877"))
}

