% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.standard.sizes}
\alias{get.standard.sizes}
\title{Function to size HPO nodes in plot 0.75 inches across}
\usage{
get.standard.sizes(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector of sizes, named by term
}
\description{
Function to size HPO nodes in plot 0.75 inches across
}

