% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sets.R
\name{get.term.descendancy.matrix}
\alias{get.term.descendancy.matrix}
\title{Get logical descendancy matrix for set of terms}
\usage{
get.term.descendancy.matrix(hpo.terms, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}
}
\value{
A logical descendancy matrix of \code{terms} by \code{terms} based on DAG structure of HPO, where by the row  term is an ancestor of the column term if result[row.term,col.term] == TRUE
}
\description{
Get logical descendancy matrix for set of terms
}
\examples{
data(hpo.terms)
get.term.descendancy.matrix(hpo.terms, c("HP:0001873", "HP:0011877"))
}

