% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.white.nodes}
\alias{get.white.nodes}
\title{Function to colour HPO nodes in plot white}
\usage{
get.white.nodes(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector of sizes, named by term
}
\description{
Function to colour HPO nodes in plot white
}

