% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo.terms}
\alias{hpo.terms}
\title{HPO Terms object (based on version 887 of the HPO)}
\format{List of indices containing metadata and structure of HPO}
\description{
HPO Terms object (based on version 887 of the HPO)
}

