\name{hpoPlot}
\docType{package}
\title{
	Functions for Plotting HPO Terms	
}
\description{
Functions for performaing operations on sets of HPO terms (character vectors of HPO term IDs) in the context of the HPO structure, and plotting them with a various options
}
\details{
\tabular{ll}{
Package: \tab hpoPlot\cr
Type: \tab Package\cr
Version: \tab 2.3\cr
Date: \tab 2015-01-7\cr
License: \tab GPL (>= 2)\cr
}
The key function is \code{\link{hpo.plot}}, which plots a set of phenotype terms given their HPO IDs and their ontological relations given by the HPO.
}
\author{
Daniel Greene <dg333@cam.ac.uk>

Maintainer: Daniel Greene <dg333@cam.ac.uk>
}
\references{
	`The Human Phenotype Ontology project: linking molecular biology and disease through phenotype data', Nucl. Acids Res. (1 January 2014) 42 (D1): D966-D974 doi:10.1093/nar/gkt1026
	Westbury, S. K. et al. (2015). Human Phenotype Ontology annotation and cluster analysis to un- ravel genetic defects in 707 cases with unexplained bleeding and platelet disorders. Genome Medicine. 7 (2015)
}
\keyword{ HPO }
