% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{intersection.with.branches}
\alias{intersection.with.branches}
\title{Intersect set of terms with branches of HPO}
\usage{
intersection.with.branches(hpo.terms, branch.roots, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{branch.roots}{Character vector of roots of branches you wish to intersect with}

\item{terms}{Character vector of HPO terms}
}
\value{
Character vector of terms
}
\description{
Intersect set of terms with branches of HPO
}
\examples{
data(hpo.terms)
intersection.with.branches(hpo.terms, "HP:0001872", c("HP:0001873", "HP:0011877"))
}

