\name{labelling.functions}
\alias{labelling.functions}
\docType{data}
\title{
List Of Labelling Functions
}
\description{
A list of functions which automatically generate labels for nodes representing HPO terms in plots generated by \code{\link{hpo.graph}}.
}
\format{
  The format is:
List of 5
 $ get.frequency.based.labels :function (hpo.terms, plotting.context)  
 $ get.simple.node.labels     :function (hpo.terms, plotting.context)  
 $ get.informative.node.labels:function (hpo.terms, plotting.context)  
 $ get.patient.based.labels   :function (hpo.terms, plotting.context)  
 $ get.code.node.labels       :function (hpo.terms, plotting.context)  
}
\keyword{datasets}
\keyword{labelling}
