% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpo.terms}
\alias{mpo.terms}
\title{MPO Terms object}
\format{List of indices containing metadata and structure of MPO}
\description{
Object comprising \code{list} of properties of the MPO, indexed by term ID
}

