% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpo.to.hpo}
\alias{mpo.to.hpo}
\title{Object containing data for mapping between MPO and HPO}
\format{List of HPO terms per MPO term}
\description{
Object containing data for mapping between MPO and HPO
}

