% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{p.values.for.occurrence.of.term.in.group}
\alias{p.values.for.occurrence.of.term.in.group}
\title{Get p-values for observing at least as many of each term as have been in phenotypes given information content}
\usage{
p.values.for.occurrence.of.term.in.group(hpo.terms, hpo.phenotypes, terms.freq)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{hpo.phenotypes}{List of HPO term character vectors}

\item{terms.freq}{Numeric vector of population frequencies of terms}
}
\value{
Numeric vector of log p-values named by correspnding term
}
\description{
Get p-values for observing at least as many of each term as have been in phenotypes given information content
}

