% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{remove.links}
\alias{remove.links}
\title{Remove terms with exactly one parent and child from plot}
\usage{
remove.links(hpo.terms, terms, plotting.context = NULL)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector
}
\description{
Remove terms with exactly one parent and child from plot
}
\examples{
data(hpo.terms)
remove.links(hpo.terms, c("HP:0001873"), list(hpo.phenotypes=list("HP:0001873", "HP:0001902")))
}
\seealso{
\code{\link{remove.terms.with.less.than.n.occurrences}}, \code{\link{n.most.frequent.terms}}
}

