% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{remove.non.pa.terms}
\alias{remove.non.pa.terms}
\title{Remove terms not descending from phenotypic abnormality}
\usage{
remove.non.pa.terms(hpo.terms, terms, plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector
}
\description{
Remove terms not descending from phenotypic abnormality
}
\seealso{
\code{\link{remove.terms.with.less.than.n.occurrences}}, \code{\link{n.most.frequent.terms}}
}

