% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{remove.terms.with.less.than.n.occurrences}
\alias{remove.terms.with.less.than.n.occurrences}
\title{Remove terms with less than certain number of occurrences}
\usage{
remove.terms.with.less.than.n.occurrences(hpo.terms, terms, plotting.context, n)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}

\item{n}{Integer}
}
\value{
Character vector
}
\description{
Remove terms with less than certain number of occurrences
}
\examples{
data(hpo.terms)
remove.terms.with.less.than.n.occurrences(hpo.terms, 
c("HP:0001873"), list(hpo.phenotypes=list("HP:0001873", "HP:0001902")), 2)
}
\seealso{
\code{\link{n.most.frequent.terms}}, \code{\link{remove.uninformative.for.plot}}
}

