% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{remove.uninformative.for.plot}
\alias{remove.uninformative.for.plot}
\title{Remove uninformative terms (fitting plotting filter format)}
\usage{
remove.uninformative.for.plot(hpo.terms, terms = get.ancestors(hpo.terms,
  unlist(plotting.context$hpo.phenotypes)), plotting.context)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}

\item{plotting.context}{List object with hpo.phenotypes slot for list of character vectors of terms}
}
\value{
Character vector
}
\description{
Remove uninformative terms (fitting plotting filter format)
}
\examples{
data(hpo.terms)
remove.uninformative.for.plot(hpo.terms, 
c("HP:0001873"), list(hpo.phenotypes=list("HP:0001873", "HP:0001902")))
}
\seealso{
\code{\link{remove.terms.with.less.than.n.occurrences}}, \code{\link{n.most.frequent.terms}}
}

