% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{remove.uninformative.terms}
\alias{remove.uninformative.terms}
\title{Get a minimal set of terms which can be used to partition a set of phenotypes}
\usage{
remove.uninformative.terms(hpo.terms, hpo.phenotypes)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{hpo.phenotypes}{List of HPO term character vectors}
}
\value{
Character vector of set of terms, excluding terms for which the presence of their descendants all partition the set of terms in the same way
}
\description{
Get a minimal set of terms which can be used to partition a set of phenotypes
}
\examples{
data(hpo.terms)
remove.uninformative.terms(hpo.terms, list(Patient1=c("HP:0001873")))
}

