% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDimNames}
\alias{setDimNames}
\title{setNames for arrays...}
\usage{
setDimNames(array.object, list.of.dimension.names)
}
\arguments{
\item{array.object}{Array}

\item{list.of.dimension.names}{List of character vectors with which to name each dimension of the array}
}
\value{
Named array
}
\description{
setNames for arrays...
}
\examples{
setDimNames(matrix(1:4,2,2), list(c("Cat", "Dog"), c("Name", "Weight")))
}

