% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalise words in character vector}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector
}
\description{
Capitalise words in character vector
}
\examples{
simpleCap(c("a simple test", "Another-test"))
}

