\name{size.functions}
\alias{size.functions}
\docType{data}
\title{
List Of Node-Sizing Functions 
}
\description{
A list of functions which set sizes of nodes representing HPO terms in plots generated by \code{\link{hpo.graph}}.
}
\format{
  The format is:
List of 3
 $ get.significance.based.sizes:function (hpo.terms, plotting.context)  
 $ get.frequency.based.sizes   :function (hpo.terms, plotting.context)  
 $ get.standard.sizes          :function (hpo.terms, plotting.context)  
}
\keyword{datasets}
\keyword{size}
