% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{term.set.list.from.character}
\alias{term.set.list.from.character}
\title{Get list of character vector of HPO terms, given character vector of comma separated terms}
\usage{
term.set.list.from.character(character.vector)
}
\arguments{
\item{character.vector}{Character vector of comma separated terms}
}
\value{
List of character vectors of HPO terms
}
\description{
Get list of character vector of HPO terms, given character vector of comma separated terms
}
\examples{
term.set.list.from.character(c("HP:0001873", "HP:0001902"))
}

