% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{clean.terms}
\alias{clean.terms}
\title{Remove redundant/implied terms}
\usage{
clean.terms(hpo.terms, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}
}
\value{
Character vector of HPO terms
}
\description{
Remove redundant/implied terms
}
\examples{
data(hpo.terms)
clean.terms(hpo.terms, c("HP:0001873", "HP:0001872"))
}

