% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get.ancestors}
\alias{get.ancestors}
\title{Get set of all ancestors of set of terms}
\usage{
get.ancestors(hpo.terms, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}
}
\value{
Character vector of all HPO terms which are an ancestor of at least one term in \code{terms}, including the terms themselves
}
\description{
Get set of all ancestors of set of terms
}
\examples{
data(hpo.terms)
get.ancestors(hpo.terms, c("HP:0001873", "HP:0011877"))
}
\seealso{
\code{link{get.descendants}}
}

