% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{get.hpo.terms}
\alias{get.hpo.terms}
\title{Gets HPO R-Object}
\usage{
get.hpo.terms(file)
}
\arguments{
\item{file}{File path of obo file}
}
\value{
R-Object (list) representing ontology
}
\description{
Gets HPO R-Object
}

