% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get.node.friendly.long.names}
\alias{get.node.friendly.long.names}
\title{Split up the HPO term descriptions so they fit in nodes for plot}
\usage{
get.node.friendly.long.names(hpo.terms, terms)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{terms}{Character vector of HPO terms}
}
\value{
Character vector
}
\description{
Split up the HPO term descriptions so they fit in nodes for plot
}
\examples{
data(hpo.terms)
get.node.friendly.long.names(hpo.terms, c("HP:0001873", "HP:0011877"))
}

