% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{get.ontology}
\alias{get.ontology}
\title{Get R-Object representation of ontology from obo file}
\usage{
get.ontology(file, species = "H")
}
\arguments{
\item{file}{File path of obo file}

\item{species}{Character vector - "H" for HPO, "M" for MPO}
}
\value{
R-Object (list) representing ontology
}
\description{
Get R-Object representation of ontology from obo file
}

