\name{hpo.terms}
\alias{hpo.terms}
\docType{data}
\title{
HPO R-Object
}
\description{
An R-object representing the HPO, and a required argument to many functions in this package.
}
\usage{data(hpo)}
\format{
	See \code{\link{get.hpo.terms}}.
}
\examples{
data(hpo)
}
\keyword{HPO}
\keyword{datasets}

