\name{hpoPlot}
\docType{package}
\title{
	Functions for Plotting HPO Terms	
}
\description{
Collection of functions for manipulating sets of HPO terms and plotting them with a various options
}
\details{
\tabular{ll}{
Package: \tab hpoPlot\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2015-01-7\cr
License: \tab GPL (>= 2)\cr
}
Key functions include set operations for sets of HPO terms represented as character vectors and hpo.plot, for plotting sets of HPO terms with respect to collections of HPO encoded phenotypes
}
\author{
Daniel Greene <dg333@cam.ac.uk>

Maintainer: Daniel Greene <dg333@cam.ac.uk>
}
\references{
	`The Human Phenotype Ontology project: linking molecular biology and disease through phenotype data', Nucl. Acids Res. (1 January 2014) 42 (D1): D966-D974 doi:10.1093/nar/gkt1026
}
\keyword{ HPO }
